fix: 예약 가격 계약과 P0 확정 경계를 보완한다

+++ src/main/java/io/regionevent/regioneventbackend/domain/content/dto/CreateContentRequest.java
@@ -48,7 +48,7 @@ public record CreateContentRequest(
 
     @NotNull
     @PositiveOrZero
-    Integer reservationPrice,
+    Long reservationPrice,
 
     @NotNull
     @JsonFormat(without = JsonFormat.Feature.ADJUST_DATES_TO_CONTEXT_TIME_ZONE)
@@ -77,7 +77,7 @@ public record CreateContentRequest(
     ) {
         this(
             title, description, locationText, operatingHoursText, contactText, precautions,
-            ageRequirement, materials, cancellationPolicyText, 0, publishAt,
+            ageRequirement, materials, cancellationPolicyText, 0L, publishAt,
             representativeImageObjectId, sessions
         );
     }
