+++ src/main/java/io/regionevent/regioneventbackend/domain/content/service/CreateContentRevisionUseCase.java
@@ -223,7 +223,7 @@ public class CreateContentRevisionUseCase {
         ));
     }
 
-    private void validateReservationPrice(Integer reservationPrice) {
+    private void validateReservationPrice(Long reservationPrice) {
         if (reservationPrice == null || reservationPrice < 0) {
             throw invalidInput();
         }
