+++ src/main/java/io/regionevent/regioneventbackend/domain/content/service/CreateContentUseCase.java
@@ -173,7 +173,7 @@ public class CreateContentUseCase {
         }
     }
 
-    private void validateReservationPrice(Integer reservationPrice) {
+    private void validateReservationPrice(Long reservationPrice) {
         if (reservationPrice == null || reservationPrice < 0) {
             throw invalidInput();
         }
