+++ src/main/java/io/regionevent/regioneventbackend/domain/content/service/UpdateMyContentUseCase.java
@@ -148,7 +148,7 @@ public class UpdateMyContentUseCase {
         );
     }
 
-    private void validateReservationPrice(Integer reservationPrice) {
+    private void validateReservationPrice(Long reservationPrice) {
         if (reservationPrice == null || reservationPrice < 0) {
             throw invalidInput();
         }
